core: Extract Coulomb communication
