+++ src/core/electrostatics_magnetostatics/elc.cpp
@@ -30,6 +30,7 @@
 #include <cmath>
 #include <mpi.h>
 
+#include "electrostatics_magnetostatics/common.hpp"
 #include "electrostatics_magnetostatics/coulomb.hpp"
 #include "electrostatics_magnetostatics/elc.hpp"
 #include "electrostatics_magnetostatics/p3m.hpp"
