+++ src/python/espressomd/electrostatics.pxd
@@ -34,7 +34,7 @@ cdef extern from "EspressoSystemInterface.hpp":
 
 
 IF ELECTROSTATICS:
-    cdef extern from "communication.hpp":
+    cdef extern from "electrostatics_magnetostatics/common.hpp":
         void mpi_bcast_coulomb_params()
 
     IF P3M:
